+++ components/List.tsx
@@ -2,7 +2,7 @@ import { getAllPosts } from '../lib/notion/getData'
 import { getAllTagsFromPosts } from '../lib/notion/getAllTagsFromPosts'
 import { setDataToCache, getDataFromCache } from '../lib/cache'
 import styles from '../components/list.module.css'
-import Pagination from '../components/Pagination'
+import Pagination from './ui/Pagination'
 
 import Time from '../components/Time'
 import abc from '../public/social.jpg'
@@ -42,12 +42,7 @@ export default async function List() {
             {list.type == "精选" ?
                <svg xmlns="http://www.w3.org/2000/svg" fill="none" version="1.1" width="16.000003814697266" height="22.30051803588867" viewBox="0 0 35.000003814697266 47.30051803588867">
                   <path d="M0,45.9367L0,0L35,0L35,45.9367C35,47.1162,33.6172,47.7374,32.7517,46.9469L18.4022,33.8438C17.8898,33.3758,17.1102,33.3758,16.5978,33.8438L2.2484,46.9469C1.38265,47.7374,0,47.1162,0,45.9367Z" fill="#838383" fill-opacity="0.27000001072883606" /> </svg>
-               : <h1>|</h1>}
-
-
-
-
-
+               : <span></span>}
 
          </div>
          <div className={styles.posts_cover}>
@@ -120,13 +115,13 @@ export default async function List() {
          <div className="mt-5 flex w-full justify-center">
             {/* <Pagination totalPages={1} /> */}
             <Pagination
-
                items={posts.length} // 100
                currentPage={1} // 1
                pageSize={10} // 10
                onPageChange={1}
 
-            /></div>
+            />
+         </div>
       </div>
    )
